Refactoring PathFinder
Elite: Dangerous has too many stars now! My crappy path finder is too
slow now. Also, this seems pretty obsolete because of the inbuilt Route
Planner in the Galaxy Map, but it's still fun working on the algorithm
so I'll keep going. Adding vector projections to see which routes
actually go to the target, but still need to get a recursive version
going to remove so many unnecessary iterations.